Extracting test utils module.
