+++ tests/conversions.rs
@@ -1,95 +1,7 @@
-use astro_coords::{
-    cartesian::Cartesian, direction::Direction, earth_equatorial::EarthEquatorial,
-    ecliptic::Ecliptic, equatorial::Equatorial, spherical::Spherical,
-};
-use simple_si_units::{base::Distance, geometry::Angle};
+mod utils;
 
-const ACC: f64 = 1e-5;
-const DISTANCE_ACC: Distance<f64> = Distance { m: ACC };
-const ANGLE_ACC: Angle<f64> = Angle { rad: ACC };
-
-fn distance_examples() -> Vec<Distance<f64>> {
-    vec![
-        Distance::from_m(0.1),
-        Distance::from_m(2.0),
-        Distance::from_m(-3.0),
-    ]
-}
-
-fn angle_examples() -> Vec<Angle<f64>> {
-    vec![
-        Angle::from_degrees(0.0),
-        Angle::from_degrees(0.1),
-        Angle::from_degrees(45.0),
-        Angle::from_degrees(90.0),
-        Angle::from_degrees(180.0),
-        Angle::from_degrees(270.0),
-        Angle::from_degrees(360.0),
-    ]
-}
-
-fn spherical_examples() -> Vec<Spherical> {
-    let mut examples = vec![];
-    for lon in angle_examples() {
-        for lat in angle_examples() {
-            examples.push(Spherical::new(lon, lat));
-        }
-    }
-    examples
-}
-
-fn cartesian_examples() -> Vec<Cartesian> {
-    let mut examples = vec![];
-    for x in distance_examples() {
-        for y in distance_examples() {
-            for z in distance_examples() {
-                examples.push(Cartesian::new(x, y, z));
-            }
-        }
-    }
-    examples
-}
-
-fn direction_examples() -> Vec<Direction> {
-    let mut examples = vec![];
-    for x in distance_examples() {
-        for y in distance_examples() {
-            for z in distance_examples() {
-                let dir = Direction::new(x.m, y.m, z.m).unwrap();
-                examples.push(dir);
-            }
-        }
-    }
-    examples
-}
-
-fn earth_equatorial_examples() -> Vec<EarthEquatorial> {
-    let mut examples = vec![];
-    for ra in angle_examples() {
-        for dec in angle_examples() {
-            examples.push(EarthEquatorial::new(ra, dec));
-        }
-    }
-    examples
-}
-
-fn ecliptic_examples() -> Vec<Ecliptic> {
-    let mut examples = vec![];
-    for spherical in spherical_examples() {
-        examples.push(Ecliptic::new(spherical));
-    }
-    examples
-}
-
-fn equatorial_examples() -> Vec<Equatorial> {
-    let mut examples = vec![];
-    for spherical in spherical_examples() {
-        for axis in direction_examples() {
-            examples.push(Equatorial::new(spherical, axis));
-        }
-    }
-    examples
-}
+use simple_si_units::base::Distance;
+use utils::{constants::*, examples::*};
 
 #[test]
 fn cartesian_to_direction_roundtrip() {
